fix: funder create deal repair
