+++ package.json
@@ -10,7 +10,7 @@
     "env:check": "tsx scripts/check-prod-env.mjs",
     "schema:check": "node scripts/check-schema-drift.mjs",
     "webhook:test": "tsx scripts/send-test-partner-webhook.ts",
-    "test": "tsx tests/release-gate.test.ts && tsx tests/docusign-webhook-hmac.test.ts && tsx tests/stripe-webhook-security.test.ts && tsx tests/partner-scope-isolation.test.ts && tsx tests/demo-reset-safety.test.ts && tsx tests/ai-review-consistency.test.ts && tsx tests/demo-canonical-data.test.ts && tsx tests/demo-contractor-flow.test.ts && tsx tests/demo-activity-feed.test.ts && tsx tests/demo-fresh-state.test.ts && tsx tests/audit-p0-coverage.test.ts && tsx tests/audit-p1-auth-onboarding.test.ts && tsx tests/audit-p1-recon-cron.test.ts && tsx tests/contrast-vektrum-blue.test.ts && tsx tests/route-smoke.test.ts && tsx tests/admin-safety.test.ts && tsx tests/booking-cta.test.ts && tsx tests/rls-regression.test.ts && tsx tests/env-validation.test.ts && tsx tests/partner-webhook-test-event.test.ts && tsx tests/audit-chain-health.test.ts && tsx tests/external-releases-ui.test.ts && tsx tests/ops-routes-safety.test.ts && tsx tests/change-orders-ui.test.ts && tsx tests/retainage-math.test.ts && tsx tests/evidence-documents-ui.test.ts && tsx tests/demo-dispute-resolution.test.ts && tsx tests/demo-activity-log.test.ts && tsx tests/demo-banner-sticky.test.ts && tsx tests/contractor-deal-authority.test.ts && tsx tests/external-rail-billing.test.ts && tsx tests/demo-walkthrough.test.ts && tsx tests/faq-demo-script.test.ts && tsx tests/schema-drift.test.ts && tsx tests/admin-partner-lifecycle.test.ts && tsx tests/deal-audit-packet.test.ts && tsx tests/admin-reconciliation-ui.test.ts && tsx tests/notifications-foundation.test.ts && tsx tests/stripe-connect-status.test.ts && tsx tests/invite-token-validation.test.ts && tsx tests/invite-role-persistence.test.ts && tsx tests/invite-public-access.test.ts && tsx tests/perplexity-draw-brief.test.ts && tsx tests/invite-funder-signup.test.ts && tsx tests/notifications-coverage.test.ts && tsx tests/milestone-documents-upload.test.ts && tsx tests/audit-pgcrypto-schema.test.ts && tsx tests/funder-mfa-onboarding.test.ts && tsx tests/deal-control-center.test.ts && tsx tests/lien-waiver-signed-url.test.ts && tsx tests/contractor-funder-display.test.ts && tsx tests/sov-foundation.test.ts && tsx tests/sov-contract-flow.test.ts && tsx tests/contract-upload-flow.test.ts && tsx tests/contractor-workflow-ia.test.ts && tsx tests/sov-milestone-links.test.ts && tsx tests/notifications-ui.test.ts && tsx tests/demo-live-product-story.test.ts && tsx tests/docusign-contract-surface.test.ts && tsx tests/demo-live-story.test.ts && tsx tests/homepage-alignment.test.ts && tsx tests/pitch-alignment.test.ts && tsx tests/ops-release-health-clarity.test.ts && tsx tests/seo-accessibility-audit.test.ts && tsx tests/marketing-cache-architecture.test.ts && tsx tests/audit-log-admin.test.ts && tsx tests/engagement-cta.test.ts && tsx tests/create-deal-flow.test.ts && tsx tests/funder-dashboard-deal-visibility.test.ts && tsx tests/docusign-envelope-notification.test.ts"
+    "test": "tsx tests/release-gate.test.ts && tsx tests/docusign-webhook-hmac.test.ts && tsx tests/stripe-webhook-security.test.ts && tsx tests/partner-scope-isolation.test.ts && tsx tests/demo-reset-safety.test.ts && tsx tests/ai-review-consistency.test.ts && tsx tests/demo-canonical-data.test.ts && tsx tests/demo-contractor-flow.test.ts && tsx tests/demo-activity-feed.test.ts && tsx tests/demo-fresh-state.test.ts && tsx tests/audit-p0-coverage.test.ts && tsx tests/audit-p1-auth-onboarding.test.ts && tsx tests/audit-p1-recon-cron.test.ts && tsx tests/contrast-vektrum-blue.test.ts && tsx tests/route-smoke.test.ts && tsx tests/admin-safety.test.ts && tsx tests/booking-cta.test.ts && tsx tests/rls-regression.test.ts && tsx tests/env-validation.test.ts && tsx tests/partner-webhook-test-event.test.ts && tsx tests/audit-chain-health.test.ts && tsx tests/external-releases-ui.test.ts && tsx tests/ops-routes-safety.test.ts && tsx tests/change-orders-ui.test.ts && tsx tests/retainage-math.test.ts && tsx tests/evidence-documents-ui.test.ts && tsx tests/demo-dispute-resolution.test.ts && tsx tests/demo-activity-log.test.ts && tsx tests/demo-banner-sticky.test.ts && tsx tests/contractor-deal-authority.test.ts && tsx tests/external-rail-billing.test.ts && tsx tests/demo-walkthrough.test.ts && tsx tests/faq-demo-script.test.ts && tsx tests/schema-drift.test.ts && tsx tests/admin-partner-lifecycle.test.ts && tsx tests/deal-audit-packet.test.ts && tsx tests/admin-reconciliation-ui.test.ts && tsx tests/notifications-foundation.test.ts && tsx tests/stripe-connect-status.test.ts && tsx tests/invite-token-validation.test.ts && tsx tests/invite-role-persistence.test.ts && tsx tests/invite-public-access.test.ts && tsx tests/perplexity-draw-brief.test.ts && tsx tests/invite-funder-signup.test.ts && tsx tests/notifications-coverage.test.ts && tsx tests/milestone-documents-upload.test.ts && tsx tests/audit-pgcrypto-schema.test.ts && tsx tests/funder-mfa-onboarding.test.ts && tsx tests/deal-control-center.test.ts && tsx tests/lien-waiver-signed-url.test.ts && tsx tests/contractor-funder-display.test.ts && tsx tests/sov-foundation.test.ts && tsx tests/sov-contract-flow.test.ts && tsx tests/contract-upload-flow.test.ts && tsx tests/contractor-workflow-ia.test.ts && tsx tests/sov-milestone-links.test.ts && tsx tests/notifications-ui.test.ts && tsx tests/demo-live-product-story.test.ts && tsx tests/docusign-contract-surface.test.ts && tsx tests/demo-live-story.test.ts && tsx tests/homepage-alignment.test.ts && tsx tests/pitch-alignment.test.ts && tsx tests/ops-release-health-clarity.test.ts && tsx tests/seo-accessibility-audit.test.ts && tsx tests/marketing-cache-architecture.test.ts && tsx tests/audit-log-admin.test.ts && tsx tests/engagement-cta.test.ts && tsx tests/create-deal-flow.test.ts && tsx tests/funder-dashboard-deal-visibility.test.ts && tsx tests/docusign-envelope-notification.test.ts && tsx tests/dashboard-role-routing.test.ts && tsx tests/new-deal-role-flash.test.ts"
   },
   "dependencies": {
     "@supabase/ssr": "^0.5.0",
